connect signup to hook
